Update README.md
In this project, we device a machine learning model to improve the process of identification of potential blood donors in a hospital. The motivation behind this work is the situation in blood banks in many small and medium sized hospitals and clinics where the data is manually entered and records are physically checked to arrive at a list of potential donors. In case of emergencies, when blood is not available readily in the bank, or when large quantity of blood is available at a short notice such as in accidents, speed of action is of utmost importance. In hospitals without a blood bank management system, the supporting staff contact potential blood donors by telephone or social media, and there is a delay between the demand and supply of blood. In this project, we use an algorithm to predict who could be the most likely candidates to donate blood, so that fewer number of people may be directly contacted, and valuable time saved. 

Logistic Regression and support vector machine shows more consistent values in train and test datasets and are more suitable prediction model. KNN and Random Forest regressor has low false negative cases. We plan to use Logistic Regression and Support Vector machine algorithm for our project.